+++ mod3rd/admin_esp/__app__.py
@@ -8,6 +8,7 @@ class WLAN_chooser(Plugin):
     def __init__(self):
         super().__init__()
         self.caption = "WLAN"
+        ## todo remove
         self.path_spec = "mod3rd.admin_esp"
         self.generators = [wlan_router]
         self.url_caption_tuple_list = [(wlan_router.root + "/wlan", None)]
@@ -17,6 +18,7 @@ class SoftAp_config(Plugin):
     def __init__(self):
         super().__init__()
         self.caption = "SoftAP"
+        ## todo remove
         self.path_spec = "mod3rd.admin_esp"
         self.generators = [softap_router]
         self.url_caption_tuple_list = [(softap_router.root + "/softap", None)]